chore: refactoring job options and job creation
